Refactored New Post page to use new Form component
