+++ app/javascript/src/components/Posts/New.jsx
@@ -1,24 +1,13 @@
 import React, { useRef, useState } from "react";
 
 import { Header, PageLoader, Sidebar } from "components/commons";
-import {
-  useCreateCategory,
-  useFetchCategories,
-} from "hooks/reactQuery/useCategories";
 import { useCreatePost } from "hooks/reactQuery/usePosts";
 import { ActionDropdown, Button } from "neetoui";
-import { Form, Input, Textarea, Select } from "neetoui/formik";
 import { useTranslation } from "react-i18next";
 import { useHistory } from "react-router-dom";
 import routes from "routes";
 
-import {
-  MAX_DESCRIPTION_LENGTH,
-  MAX_TITLE_LENGTH,
-  NEW_POST_INITIAL_VALUES,
-  NEW_POST_VALIDATION_SCHEMA,
-} from "./constants";
-import { getCategoryOptions } from "./utils";
+import NewPostForm from "./Form";
 
 const {
   Menu,
@@ -32,14 +21,7 @@ const New = () => {
 
   const formikRef = useRef();
 
-  const { mutate: createPost, isLoading: isCreatePostLoading } =
-    useCreatePost();
-
-  const { mutate: createCategory, isLoading: isCreateCategoryLoading } =
-    useCreateCategory();
-
-  const { data: { categories } = {}, isLoading: isFetchCategoriesLoading } =
-    useFetchCategories();
+  const { mutate: createPost, isLoading } = useCreatePost();
 
   const history = useHistory();
 
@@ -61,13 +43,7 @@ const New = () => {
     if (formikRef.current) formikRef.current.submitForm();
   };
 
-  const handleCreateCategory = name => {
-    createCategory({ name });
-  };
-
-  if (isFetchCategoriesLoading || isCreateCategoryLoading) {
-    return <PageLoader />;
-  }
+  if (isLoading) return <PageLoader />;
 
   return (
     <div className="flex h-screen w-screen">
@@ -78,7 +54,7 @@ const New = () => {
           actionBlock={
             <div className="ml-auto mt-auto space-x-2">
               <Button
-                disabled={isCreatePostLoading}
+                disabled={isLoading}
                 style="secondary"
                 to={routes.posts.index}
               >
@@ -101,41 +77,7 @@ const New = () => {
           }
         />
         <div className="w-full flex-1 px-16 pb-10">
-          <Form
-            className="flex h-full flex-col justify-start rounded-lg border p-10"
-            formikProps={{
-              initialValues: NEW_POST_INITIAL_VALUES,
-              validationSchema: NEW_POST_VALIDATION_SCHEMA,
-              innerRef: formikRef,
-              onSubmit: handleSubmit,
-            }}
-          >
-            <div className="space-y-4">
-              <Input
-                required
-                label={t("labels.title")}
-                maxLength={MAX_TITLE_LENGTH}
-                name="title"
-                placeholder={t("placeholders.title")}
-              />
-              <Select
-                isCreateable
-                isMulti
-                required
-                label={t("labels.category")}
-                name="categories"
-                options={getCategoryOptions(categories)}
-                onCreateOption={handleCreateCategory}
-              />
-              <Textarea
-                required
-                label={t("labels.description")}
-                maxLength={MAX_DESCRIPTION_LENGTH}
-                name="description"
-                placeholder={t("placeholders.description")}
-              />
-            </div>
-          </Form>
+          <NewPostForm innerRef={formikRef} onSubmit={handleSubmit} />
         </div>
       </div>
     </div>
